Add missing packages for aseprite
It now compiles right by hand. Compiling skia is tedious because of
reaching out to googlesource for the depot_tools set. While I was able
to run the aseprite build from this, the ansible playbook is still
missing a few steps to get this in the right place. It needs to copy the
entire bin directory to /opt/aseprite, at least.